Refactor: Split Goban functionailty into functionality layers
Split the monolithic Goban into interactive functionality,
ogs socket connectivity, and rendering base functionality.
